Add Dockerfile to have an easy build env
Facette has many dependencies and it was hard to get the correct node
version just to compile.  Adding a Dockerfile makes this much easier to
build and to run facette.

To buid:  docker build -t facette .

To run with a dynamic port: docker run -dP facette
To run with a static port of 80: docker run -d -p 80:12003 facette

I also had to update the example json file to use.